fix(theme): ship the @mui/material/styles augmentation in dist
The augmentation lived in the ambient theme-type.d.ts, which the TypeScript
program never emits, so declare module "@mui/material/styles" was absent from
dist/export.d.ts and consumers lost palette.tertiary. Move it alongside
ExtendedTheme in the compiled module; avatar-overrides.ts is the precedent for
an augmentation that does reach dist.

ExtraTheme no longer extends Theme, which removes the circularity the ambient
file carried, so theme-type.d.ts and its tsconfig types entry are both gone.
